+++ library/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Fix prototype pollution vulnerability of `record` and `recordAsync` (pull request #67)
-- Add `finite` and `safeInteger` validation (pull request #64, #65)
+- Add `finite`, `safeInteger` and `multipleOf` validation (pull request #64, #65, #66)
 
 ## v0.10.0 (August 05, 2023)